misc: move internal components into internal
