+++ src/views/Run/internal/ResultDisplay/internal/StatCard/StatCard.tsx
