+++ src/views/Run/internal/ResultDisplay/internal/StatCard/internal/StatCard.types.ts
