+++ src/views/Run/internal/ResultDisplay/internal/components.ts
@@ -1,1 +1,2 @@
 export { RunErrorDisplay } from './RunErrorDisplay'
+export { StatCard } from './StatCard'
